backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
websockets==12.0
mediapipe==0.10.11
opencv-python==4.9.0.80
numpy==1.26.4
motor==3.3.2
python-dotenv==1.0.1
